requirements.txt

Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy>=2.0.0
email-validator>=2.2.0
flask-dance>=7.1.0
flask-login>=0.6.3
oauthlib>=3.2.2
pyjwt>=2.10.1
gunicorn>=23.0.0
